backend/requirements.txt
fastapi
uvicorn
pydantic
crewai
python-dotenv
feedparser
crewai_tools